requirements.txt
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
filelock==3.15.4
fpdf==1.7.2
fsspec==2024.6.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
jiter==0.5.0
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
